chore: Split beta implementation into modules
