+++ src/function/beta/api.rs
@@ -1,0 +1,62 @@
+use super::*;
+
+/// Computes the beta function
+/// where `a` is the first beta parameter
+/// and `b` is the second beta parameter.
+///
+///
+/// # Panics
+///
+/// if `a <= 0.0` or `b <= 0.0`
+pub fn beta(a: f64, b: f64) -> f64 {
+    checked_beta(a, b).unwrap()
+}
+
+/// Computes the beta function
+/// where `a` is the first beta parameter
+/// and `b` is the second beta parameter.
+///
+///
+/// # Errors
+///
+/// if `a <= 0.0` or `b <= 0.0`
+pub fn checked_beta(a: f64, b: f64) -> Result<f64, BetaFuncError> {
+    checked_ln_beta(a, b).map(|x| x.exp())
+}
+
+/// Computes the lower incomplete (unregularized) beta function
+/// `B(a,b,x) = int(t^(a-1)*(1-t)^(b-1),t=0..x)` for `a > 0, b > 0, 1 >= x >= 0`
+/// where `a` is the first beta parameter, `b` is the second beta parameter, and
+/// `x` is the upper limit of the integral
+///
+/// # Panics
+///
+/// If `a <= 0.0`, `b <= 0.0`, `x < 0.0`, or `x > 1.0`
+pub fn beta_inc(a: f64, b: f64, x: f64) -> f64 {
+    checked_beta_inc(a, b, x).unwrap()
+}
+
+/// Computes the lower incomplete (unregularized) beta function
+/// `B(a,b,x) = int(t^(a-1)*(1-t)^(b-1),t=0..x)` for `a > 0, b > 0, 1 >= x >= 0`
+/// where `a` is the first beta parameter, `b` is the second beta parameter, and
+/// `x` is the upper limit of the integral
+///
+/// # Errors
+///
+/// If `a <= 0.0`, `b <= 0.0`, `x < 0.0`, or `x > 1.0`
+pub fn checked_beta_inc(a: f64, b: f64, x: f64) -> Result<f64, BetaFuncError> {
+    checked_beta_reg(a, b, x).and_then(|x| checked_beta(a, b).map(|y| x * y))
+}
+
+/// Computes the regularized lower incomplete beta function
+/// `I_x(a,b) = 1/Beta(a,b) * int(t^(a-1)*(1-t)^(b-1), t=0..x)`
+/// `a > 0`, `b > 0`, `1 >= x >= 0` where `a` is the first beta parameter,
+/// `b` is the second beta parameter, and `x` is the upper limit of the
+/// integral.
+///
+/// # Panics
+///
+/// if `a <= 0.0`, `b <= 0.0`, `x < 0.0`, or `x > 1.0`
+pub fn beta_reg(a: f64, b: f64, x: f64) -> f64 {
+    checked_beta_reg(a, b, x).unwrap()
+}
